+++ server/db/migrations/sqlite/0006_tan_unus.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `paperless_documents` ADD `deleted_at` integer;